feat: remove bootstrap and createApp API
