--- packages/document/main-doc/docs/zh/apis/app/runtime/core/create-app.mdx
@@ -1,35 +1,0 @@
----
-title: createApp
----
-# createApp
-
-:::warning
-即将废弃，推荐使用 [`createRoot`](/apis/app/runtime/core/create-root)。
-:::
-
-用于创建自定义入口，定制运行时插件。
-
-## 使用姿势
-
-```ts
-import { createApp } from '@modern-js/runtime';
-```
-
-## 函数签名
-
-```ts
-import type { Plugin } from '@modern-js/runtime';
-
-function createApp(options: { plugins: Plugin[] }): React.ComponentType<any>;
-```
-
-### 参数
-
-- `options`: 可选的参数。
-  - `plugins`：自定义的插件扩展。
-
-## 示例
-
-### 创建自定义入口
-
-详见 [`bootstrap`](./bootstrap.mdx)。
