+++ packages/runtime/plugin-runtime/src/core/index.ts
@@ -1,11 +1,12 @@
 export type { RuntimePlugin } from './plugin';
 export { defineConfig, getConfig, defineRuntimeConfig } from './config';
 
-// compatible
-export * from './compatible';
-
 export type { RuntimeContext } from './context/runtime';
-export { RuntimeReactContext, ServerRouterContext } from './context/runtime';
+export {
+  RuntimeReactContext,
+  ServerRouterContext,
+  useRuntimeContext,
+} from './context/runtime';
 export * from './loader';
 
 export type { SSRData, SSRContainer } from './types';
