--- packages/runtime/plugin-runtime/tests/index.test.ts
@@ -1,7 +1,0 @@
-import { createApp } from '../src';
-
-describe('plugin-runtime', () => {
-  it('default', () => {
-    expect(createApp).toBeDefined();
-  });
-});
